+++ src/main/java/com/issuetracker/engine/IssueTrackerPro.java
@@ -1,0 +1,4 @@
+package com.issuetracker.engine;
+
+public class IssueTrackerPro {
+}
