+++ src/main/java/com/issuetracker/engine/api/controller/UserController.java
@@ -1,0 +1,25 @@
+package com.issuetracker.engine.controller;
+
+import com.issuetracker.engine.model.User;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+import java.util.List;
+
+@Controller
+@RequestMapping("/users")
+public class UserController {
+
+    @Autowired //Usado para obter uma instancia
+    private User userRepository;
+
+    @GetMapping("/")
+    public ResponseEntity<List<User>> getUsers(){
+        List<User> users = userRepository.findAll();
+        return new ResponseEntity<>(users, HttpStatus.OK);
+    }
+
+}
