+++ src/main/java/com/issuetracker/engine/domain/dto/UserDto.java
@@ -1,0 +1,2 @@
+package com.issuetracker.engine.domain.dto;public class UserDto {
+}
