+++ src/main/java/com/issuetracker/engine/domain/exception/IssueException.java
