+++ src/main/java/com/issuetracker/engine/domain/repository/UserEntityRepository.java
@@ -1,0 +1,2 @@
+package com.issuetracker.engine.domain.repository;public interface UserEntityRepository {
+}
