+++ src/main/java/com/issuetracker/engine/domain/service/UserService.java
@@ -1,0 +1,2 @@
+package com.issuetracker.engine.domain.service;public class UserService {
+}
